cast concentrations to ints to avoid having '512000.0' as a key in the HDF5 file

+++ champ/misc.py
@@ -58,7 +58,7 @@ def parse_concentration(filename):
     m = re.search(pattern, filename)
     if m is None:
         raise ValueError("The concentration cannot be parsed from the filename: %s" % filename)
-    conc = float(m.group(1).replace('_', '.'))
+    conc = int(m.group(1).replace('_', '.'))
     if m.group(2).lower() == 'pm':
         return conc
     elif m.group(2).lower() == 'nm':
